requirements.txt
akshare
pandas
numpy
scikit-learn
streamlit
plotly
pyyaml
